ToxD4C_framework/requirements.txt
# ToxD4C 项目依赖
# 核心深度学习框架
torch>=2.0.0
torch-geometric>=2.6.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torch-spline-conv>=1.2.0

# 科学计算和数据处理
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 化学信息学
rdkit-pypi>=2022.3.0
mordred>=1.2.0

# 机器学习工具
scikit-learn>=1.0.0
tqdm>=4.62.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
requests>=2.25.0
aiohttp>=3.8.0
fsspec>=2022.1.0
jinja2>=3.0.0
psutil>=5.8.0
pyparsing>=3.0.0